+++ aim/web/ui/src/services/models/imagesExplore/imagesExploreAppModel.ts
@@ -46,7 +46,6 @@ import {
   ISelectConfig,
   ISelectOption,
 } from 'types/services/models/explorer/createAppModel';
-import { IApiRequest } from 'types/services/services';
 
 import onRowSelectAction from 'utils/app/onRowSelect';
 import { decode, encode } from 'utils/encoder/encoder';
@@ -738,6 +737,11 @@ function getGroupingSelectOptions({
       group: 'images',
       label: 'images.name',
       value: 'images_name',
+    },
+    {
+      group: 'images',
+      label: 'images.context',
+      value: 'context',
     },
     ...contextOptions,
     {