+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Enhancements:
+
+- Ability to group by metric.context (VkoHov)
+
 ## 3.5.1 Feb 4 2022
 
 - Fix folder creation when tracking with remote tracker (aramaim)